Refactor backend http handlers
