+++ proxy/backend.go
@@ -25,7 +25,6 @@ package proxy
 
 import (
 	"context"
-	"crypto/sha256"
 	"crypto/tls"
 	"crypto/x509"
 	"encoding/hex"
@@ -34,37 +33,12 @@ import (
 	"io"
 	"log"
 	"net"
-	"net/http"
-	"net/http/httputil"
-	"regexp"
-	"runtime/debug"
 	"slices"
-	"strconv"
-	"strings"
 	"time"
 
 	"github.com/pires/go-proxyproto"
-	"golang.org/x/net/http2"
 
 	"github.com/c2FmZQ/tlsproxy/proxy/internal/netw"
-)
-
-const (
-	hstsHeader = "Strict-Transport-Security"
-	hstsValue  = "max-age=2592000" // 30 days
-
-	viaHeader           = "Via"
-	hostHeader          = "Host"
-	xForwardedForHeader = "X-Forwarded-For"
-)
-
-type ctxURLKeyType int
-
-var (
-	ctxURLKey        ctxURLKeyType = 1
-	ctxOverrideIDKey ctxURLKeyType = 2
-
-	commaRE = regexp.MustCompile(`, *`)
 )
 
 func (be *Backend) incInFlight(delta int) int {
@@ -294,243 +268,6 @@ func (be *Backend) bridgeConns(client, server net.Conn) error {
 		retErr = fmt.Errorf("[int➔ ext]: %w", unwrapErr(err))
 	}
 	return retErr
-}
-
-func (be *Backend) localHandlersAndAuthz(next http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, req *http.Request) {
-		reqHost := hostFromReq(req)
-		reqPath := req.URL.Path
-		hi := slices.IndexFunc(be.localHandlers, func(h localHandler) bool {
-			if h.host != "" && h.host != reqHost {
-				return false
-			}
-			return h.path == reqPath || (h.matchPrefix && strings.HasPrefix(reqPath, h.path+"/"))
-		})
-		if hi >= 0 && be.localHandlers[hi].ssoBypass {
-			be.localHandlers[hi].handler.ServeHTTP(w, req)
-			return
-		}
-		if !be.enforceSSOPolicy(w, req) {
-			return
-		}
-		if hi >= 0 && !be.localHandlers[hi].ssoBypass {
-			be.localHandlers[hi].handler.ServeHTTP(w, req)
-			return
-		}
-		if hi < 0 {
-			pathSlash := reqPath + "/"
-			if hi := slices.IndexFunc(be.localHandlers, func(h localHandler) bool {
-				if h.host != "" && h.host != reqHost {
-					return false
-				}
-				return pathSlash == h.path
-			}); hi >= 0 {
-				http.Redirect(w, req, pathSlash, http.StatusMovedPermanently)
-				return
-			}
-		}
-		if next == nil {
-			http.NotFound(w, req)
-			return
-		}
-		next.ServeHTTP(w, req)
-	})
-}
-
-func recoverHandler(next http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, req *http.Request) {
-		defer func() {
-			if r := recover(); r != nil {
-				log.Printf("PANIC: %#v\n%s", r, string(debug.Stack()))
-			}
-		}()
-		next.ServeHTTP(w, req)
-	})
-}
-
-func (be *Backend) localHandler() http.Handler {
-	h := be.userAuthentication(be.localHandlersAndAuthz(nil))
-	return recoverHandler(http.HandlerFunc(func(w http.ResponseWriter, req *http.Request) {
-		be.setAltSvc(w.Header(), req)
-		h.ServeHTTP(w, req)
-	}))
-}
-
-func (be *Backend) setAltSvc(header http.Header, req *http.Request) {
-	if be.http3Handler == nil {
-		return
-	}
-	if req.TLS != nil && req.TLS.NegotiatedProtocol == "h3" {
-		return
-	}
-	if be.ALPNProtos == nil || !slices.Contains(*be.ALPNProtos, "h3") {
-		return
-	}
-	_, port, _ := net.SplitHostPort(req.Host)
-	if port == "" {
-		port = "443"
-	}
-	if p, err := strconv.Atoi(port); err == nil && p > 0 && p < 65536 {
-		header.Set("Alt-Svc", fmt.Sprintf("h3=\":%d\"; ma=2592000;", p))
-	}
-}
-
-func (be *Backend) reverseProxy() http.Handler {
-	if len(be.Addresses) == 0 {
-		h := http.HandlerFunc(func(w http.ResponseWriter, req *http.Request) {
-			log.Printf("PRX %s ➔ %s %s ➔ status:%d (%q)", formatReqDesc(req), req.Method, req.URL, http.StatusNotFound, userAgent(req))
-			http.NotFound(w, req)
-		})
-		return recoverHandler(be.userAuthentication(be.localHandlersAndAuthz(h)))
-	}
-	reverseProxy := &httputil.ReverseProxy{
-		Director:       be.reverseProxyDirector,
-		Transport:      be.reverseProxyTransport(),
-		ModifyResponse: be.reverseProxyModifyResponse,
-	}
-
-	return recoverHandler(be.userAuthentication(be.localHandlersAndAuthz(http.HandlerFunc(func(w http.ResponseWriter, req *http.Request) {
-		ctx := req.Context()
-		host := req.Host
-		if host == "" {
-			host = connServerName(ctx.Value(connCtxKey).(net.Conn))
-		}
-		req.URL.Scheme = "https"
-		if be.Mode == ModeHTTP {
-			req.URL.Scheme = "http"
-		}
-		req.URL.Host = host
-		req.Header.Set(hostHeader, host)
-		hostname := req.URL.Hostname()
-
-		if !slices.Contains(be.ServerNames, hostname) {
-			if req.Body != nil {
-				req.Body.Close()
-			}
-			http.Error(w, "Misdirected Request", http.StatusMisdirectedRequest)
-			return
-		}
-
-		ctx = context.WithValue(ctx, ctxURLKey, req.URL.String())
-
-		h := sha256.Sum256([]byte(hostname))
-		hh := hex.EncodeToString(h[:])
-		req.URL.Host = hh
-	L:
-		for i, po := range be.PathOverrides {
-			for _, prefix := range po.Paths {
-				if !strings.HasPrefix(req.URL.Path, prefix) {
-					continue
-				}
-				req.URL.Host = fmt.Sprintf("%s-%d", hh, i)
-				ctx = context.WithValue(ctx, ctxOverrideIDKey, i)
-				break L
-			}
-		}
-		req = req.WithContext(ctx)
-
-		hops := commaRE.Split(req.Header.Get(viaHeader), -1)
-		_, me, _ := strings.Cut(hops[len(hops)-1], " ")
-		hops = hops[:len(hops)-1]
-		for _, via := range hops {
-			_, via, _ = strings.Cut(via, " ")
-			if via != me {
-				continue
-			}
-			if req.Body != nil {
-				req.Body.Close()
-			}
-			http.Error(w, req.Header.Get(viaHeader), http.StatusLoopDetected)
-			return
-		}
-		reverseProxy.ServeHTTP(w, req)
-	}))))
-}
-
-func (be *Backend) reverseProxyDirector(req *http.Request) {
-	req.Header.Del(xForwardedForHeader)
-	req.Header.Del(xFCCHeader)
-
-	var via []string
-	if v := req.Header.Get(viaHeader); v != "" {
-		via = commaRE.Split(req.Header.Get(viaHeader), -1)
-	}
-	via = append(via, req.Proto+" "+req.Context().Value(connCtxKey).(net.Conn).LocalAddr().String())
-	req.Header.Set(viaHeader, strings.Join(via, ", "))
-
-	if req.TLS != nil && len(req.TLS.PeerCertificates) > 0 && be.ClientAuth != nil && len(be.ClientAuth.AddClientCertHeader) > 0 {
-		addXFCCHeader(req, be.ClientAuth.AddClientCertHeader)
-	}
-}
-
-type funcRoundTripper func(req *http.Request) (*http.Response, error)
-
-func (rt funcRoundTripper) RoundTrip(req *http.Request) (*http.Response, error) {
-	return rt(req)
-}
-
-func (be *Backend) reverseProxyTransport() http.RoundTripper {
-	h1 := &http.Transport{
-		DialContext: func(ctx context.Context, network, addr string) (net.Conn, error) {
-			return be.dial(ctx)
-		},
-		DialTLSContext: func(ctx context.Context, network, addr string) (net.Conn, error) {
-			return be.dial(ctx)
-		},
-		MaxIdleConns:          100,
-		IdleConnTimeout:       30 * time.Second,
-		ExpectContinueTimeout: 1 * time.Second,
-	}
-	h2 := &http2.Transport{
-		DialTLSContext: func(ctx context.Context, network, addr string, _ *tls.Config) (net.Conn, error) {
-			return be.dial(ctx, "h2")
-		},
-		DisableCompression: true,
-		AllowHTTP:          true,
-		ReadIdleTimeout:    30 * time.Second,
-		WriteByteTimeout:   30 * time.Second,
-		CountError: func(errType string) {
-			be.recordEvent("http2 client error: " + errType)
-		},
-	}
-	h3 := be.http3Transport()
-
-	return funcRoundTripper(func(req *http.Request) (*http.Response, error) {
-		proto := "http/1.1"
-		if id, ok := req.Context().Value(ctxOverrideIDKey).(int); ok && id >= 0 && id < len(be.PathOverrides) && be.PathOverrides[id].BackendProto != nil {
-			proto = *be.PathOverrides[id].BackendProto
-		} else if be.BackendProto != nil {
-			proto = *be.BackendProto
-		}
-		if proto == "" && req.TLS != nil && req.TLS.NegotiatedProtocol != "" {
-			proto = req.TLS.NegotiatedProtocol
-		}
-		if proto == "h3" && h3 != nil {
-			return h3.RoundTrip(req)
-		}
-		if proto == "h2" {
-			return h2.RoundTrip(req)
-		}
-		return h1.RoundTrip(req)
-	})
-}
-
-func (be *Backend) reverseProxyModifyResponse(resp *http.Response) error {
-	req := resp.Request
-	var cl string
-	if resp.ContentLength != -1 {
-		cl = fmt.Sprintf(" content-length:%d", resp.ContentLength)
-	}
-	url, _ := req.Context().Value(ctxURLKey).(string)
-	log.Printf("PRX %s ➔ %s %s ➔ status:%d%s (%q)", formatReqDesc(req), req.Method, url, resp.StatusCode, cl, userAgent(req))
-
-	if resp.StatusCode != http.StatusMisdirectedRequest && resp.Header.Get(hstsHeader) == "" {
-		resp.Header.Set(hstsHeader, hstsValue)
-	}
-	if resp.StatusCode >= 200 && resp.StatusCode < 400 && resp.Header.Get("Alt-Svc") == "" {
-		be.setAltSvc(resp.Header, req)
-	}
-	return nil
 }
 
 func forward(out net.Conn, in net.Conn, closeWhenDone bool, halfClosedTimeout time.Duration) error {
